Modify the type of stableId in CowDTO

+++ src/main/java/hu/flowacademy/springpracticecows/dtos/CowDTO.java
@@ -14,5 +14,5 @@ public class CowDTO {
     private Integer age;
 
     @NotEmpty
-    private Integer stableId;
+    private Long stableId;
 }
